Extract station filtering state, handlers controls
to a separate hook and component
